Fix publish when package has dependencies
When running `apm publish` on a package with a dependency, I got this problem.

```
✓ Package content verified
⣻ Checking dependency: starling-source@2.7.0
D::ProcessQueue::UNHANDLED PROCESS ERROR IN com.apm.client.commands.packages.processes::PackageDependenciesVerifyProcess
E::ProcessQueue::Error #1009: Cannot access a property or method of a null object reference.
E::ProcessQueue::TypeError: Error #1009: Cannot access a property or method of a null object reference.
        at com.apm.client.repositories::PackageResolver/getPackageVersion()
        at com.apm.client.commands.packages.processes::PackageDependenciesVerifyProcess/start()
        at com.apm.client.processes::ProcessQueue/checkAndStartNextProcess()
        at com.apm.client.processes::ProcessQueue/process_eventHandler()
        at flash.events::EventDispatcher/dispatchEventFunction()
        at flash.events::EventDispatcher/dispatchEvent()
        at com.apm.client.processes::ProcessBase/complete()
        at Function/PackageRemoteContentVerifyProcess.as$0:anonymous()
        at com.apm.client.commands.packages.processes::PackageRemoteContentVerifyProcess/loader_statusHandler()
        at flash.events::EventDispatcher/dispatchEventFunction()
        at flash.events::EventDispatcher/dispatchEvent()
        at flash.net::URLLoader/redirectEvent()
ERROR :: Error #1009: Cannot access a property or method of a null object reference.
```

`PackageResolver` is used not in the context of a project, so APM.config.projectDefinition is null.